+++ example/lib/decoration/barrel_dragable.dart
@@ -5,7 +5,8 @@ import 'package:example/util/common_sprite_sheet.dart';
 import 'package:flutter/material.dart';
 
 class BarrelDraggable extends GameDecoration with DragGesture, ObjectCollision {
-  late TextConfig _textConfig;
+  late TextPaint _textConfig;
+
   BarrelDraggable(Vector2 position)
       : super.withSprite(
           CommonSpriteSheet.barrelSprite,
@@ -26,17 +27,12 @@ class BarrelDraggable extends GameDecoration with DragGesture, ObjectCollision {
               DungeonMap.tileSize * 0.4,
             ),
           ),
-          // CollisionArea.circle(
-          //   radius: (DungeonMap.tileSize) / 2,
-          // align: Vector2(
-          //   DungeonMap.tileSize * 0.2,
-          //   DungeonMap.tileSize * 0.4,
-          // ),
-          // ),
         ],
       ),
     );
-    _textConfig = TextConfig(color: Colors.white, fontSize: width / 4);
+    _textConfig = TextPaint(
+      config: TextPaintConfig(color: Colors.white, fontSize: width / 4),
+    );
   }
 
   @override