+++ example/lib/enemy/goblin.dart
@@ -131,7 +131,7 @@ class Goblin extends SimpleEnemy with ObjectCollision {
   void receiveDamage(double damage, dynamic from) {
     this.showDamage(
       damage,
-      config: TextConfig(
+      config: TextPaintConfig(
         fontSize: width / 3,
         color: Colors.white,
       ),